feat: sidebar to component
